+++ src/internal/connector/graph_connector_helper_test.go
@@ -3,6 +3,7 @@ package connector
 import (
 	"context"
 	"io"
+	"net/http"
 	"reflect"
 	"testing"
 
@@ -335,12 +336,12 @@ func checkContact(
 	testEmptyOrEqual(t, expected.GetAssistantName(), got.GetAssistantName(), "AssistantName")
 
 	testEmptyOrEqual(t, expected.GetBirthday(), got.GetBirthday(), "Birthday")
-
-	assert.Equal(t, expected.GetBusinessAddress(), got.GetBusinessAddress())
-
-	testEmptyOrEqual(t, expected.GetBusinessHomePage(), got.GetBusinessHomePage(), "BusinessHomePage")
-
-	assert.Equal(t, expected.GetBusinessPhones(), got.GetBusinessPhones())
+	// Not present in msgraph-beta-sdk/models
+	// assert.Equal(t, expected.GetBusinessAddress(), got.GetBusinessAddress())
+	// Not present in msgraph-beta-sdk/models
+	// testEmptyOrEqual(t, expected.GetBusinessHomePage(), got.GetBusinessHomePage(), "BusinessHomePage")
+	// Not present in msgraph-beta-sdk/models
+	// assert.Equal(t, expected.GetBusinessPhones(), got.GetBusinessPhones())
 
 	assert.Equal(t, expected.GetCategories(), got.GetCategories())
 
@@ -364,9 +365,10 @@ func checkContact(
 
 	testEmptyOrEqual(t, expected.GetGivenName(), got.GetGivenName(), "GivenName")
 
-	assert.Equal(t, expected.GetHomeAddress(), got.GetHomeAddress())
-
-	assert.Equal(t, expected.GetHomePhones(), got.GetHomePhones())
+	// Not present in msgraph-beta-sdk/models
+	// assert.Equal(t, expected.GetHomeAddress(), got.GetHomeAddress())
+	// Not present in msgraph-beta-sdk/models
+	// assert.Equal(t, expected.GetHomePhones(), got.GetHomePhones())
 
 	// Skip CreatedDateTime as it's tied to this specific instance of the item.
 
@@ -382,13 +384,14 @@ func checkContact(
 
 	testEmptyOrEqual(t, expected.GetMiddleName(), got.GetMiddleName(), "MiddleName")
 
-	testEmptyOrEqual(t, expected.GetMobilePhone(), got.GetMobilePhone(), "MobilePhone")
+	// Not present in msgraph-beta-sdk/models
+	// testEmptyOrEqual(t, expected.GetMobilePhone(), got.GetMobilePhone(), "MobilePhone")
 
 	testEmptyOrEqual(t, expected.GetNickName(), got.GetNickName(), "NickName")
 
 	testEmptyOrEqual(t, expected.GetOfficeLocation(), got.GetOfficeLocation(), "OfficeLocation")
-
-	assert.Equal(t, expected.GetOtherAddress(), got.GetOtherAddress())
+	// Not present in msgraph-beta-sdk/models
+	// assert.Equal(t, expected.GetOtherAddress(), got.GetOtherAddress())
 
 	// Skip ParentFolderId as it's tied to this specific instance of the item.
 
@@ -976,9 +979,9 @@ func getSelectorWith(
 	}
 }
 
-func loadConnector(ctx context.Context, t *testing.T, r resource) *GraphConnector {
+func loadConnector(ctx context.Context, t *testing.T, itemClient *http.Client, r resource) *GraphConnector {
 	a := tester.NewM365Account(t)
-	connector, err := NewGraphConnector(ctx, a, r)
+	connector, err := NewGraphConnector(ctx, itemClient, a, r)
 	require.NoError(t, err)
 
 	return connector